refact(api/schemas): migrate system-domain local BaseModel classes to schemas_system.py (#6042)
Phase 3 of #6042: moves all local Pydantic request/response classes from
9 system-domain endpoint files into schemas_system.py.

Files migrated:
- permissions.py (11 classes: PermissionModeResponse/Request, PermissionRule*,
  PermissionAddRuleRequest, PermissionRemoveRuleRequest, ApprovalRecordResponse,
  ProjectApprovalsResponse, PermissionStatusResponse, CheckCommandRequest/Response)
- vnc_manager.py (25 classes: Vnc* responses, Mouse/Keyboard/Special/Scroll/Drag
  request classes, ConnectionQualitySettings, ConnectionSettings,
  MacroAction, MacroRecording)
- scheduler.py (19 classes: Scheduler* responses + ScheduleWorkflowRequest,
  RescheduleRequest, QueueControlRequest)
- monitoring.py (18 classes: ServicesSummaryResponse, MonitoringActionResponse,
  ClaudeApiStatusResponse, AlertCheckResponse, AlertManagerResponse, etc.)
- vnc_mcp.py (16 classes: MCPTool renamed VncMCPTool, VNCStatusRequest,
  VNCObservationRequest, VncStatusMcpResponse, Desktop* responses/requests)
- cache_management.py (13 classes: CacheStatsResponse, Cache*Response,
  Redis*Response, WarmCacheResponse, InvalidateCacheResponse, ClearAllResponse)
- browser_mcp.py (21 classes: Browser* request classes renamed from generic
  names; MCPTool now shared from schemas_code.py)
- settings.py (13 classes: SaveSettingsResponse, ClearCacheResponse,
  TaskQueuedResponse/TaskStatusResponse renamed Settings*, RBACInitRequest,
  SystemUpdateRequest, ConfigSyncRequest/Response, HardwarePriority*)
- log_forwarding.py (13 classes: DestinationCreate/Update/Response renamed
  LogFwdDestination*, LogFwdMessageResponse, LogFwdStatusResponse, etc.)

Adds to schemas_system.py: import re, datetime, timezone, Union, Field,
RetryConfig, Metadata; inlines _VALID_HARDWARE_TYPES frozenset.
